#23: implement task controller, service, and repository with TypeScript
